backend/requirements.txt
flask==3.1.0
flask-socketio==5.4.1
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
flask-migrate==4.1.0
python-dotenv==1.0.1
gtfs-realtime-bindings==1.0.0
protobuf==5.29.3
firebase-admin==6.6.0
redis==5.2.1
psycopg2-binary==2.9.10
eventlet==0.38.2
requests==2.32.3
